requirements.txt
pandas
numpy
plotly
google-auth
google-api-python-client
python-dotenv
pytz